Add expand functionality to pg_depend_recurse
Some dependencies are not recorded in pg_depend, like types depending on types. For this we want to expand the objects we traverse with extra pg_depend entries that we recurse on.

For types we expand the type to a dependency on its relation, even though the pg_depend relation is an internal dependency on the inverse we are walking the graph. This allows types depended on by the relation to be included in the graph.

TODO we probably need to expand array types to depend on their type element.